feat: Rename accountsettings -> settings
